requirements.txt
# CodeArena - 20 Day Python & MySQL Learning System
# =================================================

# Core Database
mysql-connector-python>=8.0.0

# Data Analysis (Day 17-18)
pandas>=2.0.0

# Web Framework (Day 20 CLI examples)
flask>=3.0.0

# Testing Framework
pytest>=7.0.0

# Additional Utilities
python-dateutil>=2.8.0

# Development Tools (Optional)
# black>=23.0.0          # Code formatting
# flake8>=6.0.0          # Linting
# mypy>=1.0.0            # Type checking
